// packages/client/src/store/modal/types.ts
export enum ModalType {
  SignIn = 'signin',
  SignUp = 'signup',
  RestorePassword = 'RestorePassword',
  ChatRain = 'chatRain',
  ChatTipping = 'chatTipping',
  OAuthSignup = 'oauthSignup',
  SignUpMetamask = 'signupMetamask',
  SignUpTelegram = 'signupTelegram',
  SelectDateTransaction = 'SelectDateTransaction',
  SelectDateDeposit = 'SelectDateDeposit',
  WithdrawalConfirm = 'withdrawalConfirm',
  Withdrawal = 'withdrawal',
  WithdrawalSuccess = 'withdrawalSuccess',
  SportBetsDate = 'sportBetsDate',
  SportBetsFilter = 'sportBetsFilter',
  TransactionFilter = 'transactionFilter',
  GameLiveStats = 'gameLiveStats',
  ManageCurrency = 'manageCurrency',
  DeletePhoto = 'deletePhoto',
  EditAvatar = 'editAvatar',
  DepositFilter = 'depositFilter',
  DepositDetailsModal = 'depositDetails',
  AddBonus = 'addBonus',
  CreateCampaign = 'createCampaign',
  FiatCurrency = 'fiatCurrency',
  ReferralFilter = 'ReferralFilter',
  TwoFA = 'TwoFA',
}